+++ packages/sources/tiingo/test/integration/adapter.test.ts
@@ -229,6 +229,75 @@ describe('execute', () => {
       expect(response.body).toMatchSnapshot()
     })
   })
+
+  describe('cryptoyield endpoint without poolCode', () => {
+    const data = {
+      id,
+      data: {
+        endpoint: 'cryptoyield',
+      },
+    }
+
+    it('should return success', async () => {
+      mockResponseSuccess()
+
+      const response = await (context.req as SuperTest<Test>)
+        .post('/')
+        .send(data)
+        .set('Accept', '*/*')
+        .set('Content-Type', 'application/json')
+        .expect('Content-Type', /json/)
+        .expect(200)
+      expect(response.body).toMatchSnapshot()
+    })
+  })
+
+  describe('cryptoyield endpoint with poolCode', () => {
+    const data = {
+      id,
+      data: {
+        endpoint: 'cryptoyield',
+        poolCode: 'compound_usdt',
+      },
+    }
+
+    it('should return success', async () => {
+      mockResponseSuccess()
+
+      const response = await (context.req as SuperTest<Test>)
+        .post('/')
+        .send(data)
+        .set('Accept', '*/*')
+        .set('Content-Type', 'application/json')
+        .expect('Content-Type', /json/)
+        .expect(200)
+      expect(response.body).toMatchSnapshot()
+    })
+  })
+
+  describe('realized-vol endpoint', () => {
+    const data = {
+      id,
+      data: {
+        base: 'ETH',
+        endpoint: 'realized-vol',
+      },
+    }
+
+    it('should return success', async () => {
+      mockResponseSuccess()
+
+      const response = await (context.req as SuperTest<Test>)
+        .post('/')
+        .send(data)
+        .set('Accept', '*/*')
+        .set('Content-Type', 'application/json')
+        .expect('Content-Type', /json/)
+        .expect(200)
+
+      expect(response.body).toMatchSnapshot()
+    })
+  })
 })
 
 describe('websocket', () => {